requirements.txt
ace_tools==0.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.6.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
autoray==0.7.0
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
contourpy==1.3.0
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
diastatic-malt==2.15.2
dill==0.3.9
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gradio==5.31.0
gradio_client==1.10.1
greenlet==3.1.1
groovy==0.1.2
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.32.2
ibm-cloud-sdk-core==3.22.1
ibm-platform-services==0.59.1
idna==3.10
Jinja2==3.1.6
jiter==0.6.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-text-splitters==0.3.0
langsmith==0.1.132
Levenshtein==0.26.0
llm==0.18
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.1
openpyxl==3.1.5
openqasm3==1.0.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pbr==6.1.1
PennyLane==0.40.0
PennyLane-qiskit==0.40.0
PennyLane_Lightning==0.40.0
pillow==10.4.0
pluggy==1.5.0
psutil==7.0.0
puremagic==1.28
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.1.4
pyspnego==0.11.2
python-constraint==1.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2024.2
PyYAML==6.0.2
qiskit==1.2.4
qiskit-aer==0.16.1
qiskit-ibm-provider==0.11.0
qiskit-ibm-runtime==0.29.0
qiskit-qasm3-import==0.5.1
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rich==14.0.0
ruff==0.11.11
rustworkx==0.16.0
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scipy-openblas32==0.3.29.0.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
starlette==0.46.2
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.20.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.66.5
transformers==4.45.2
triton==3.2.0
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.2
websocket-client==1.8.0
websockets==15.0
yarl==1.13.1
z3-solver==4.14.0.0
